move DrawableUtils to base-ui module
