+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/activity/TractActivity.java
@@ -43,6 +43,7 @@ import org.cru.godtools.base.model.Event;
 import org.cru.godtools.base.tool.activity.ImmersiveActivity;
 import org.cru.godtools.base.tool.model.view.ManifestViewUtils;
 import org.cru.godtools.base.tool.widget.ScaledPicassoImageView;
+import org.cru.godtools.base.ui.util.DrawableUtils;
 import org.cru.godtools.download.manager.DownloadProgress;
 import org.cru.godtools.download.manager.GodToolsDownloadManager;
 import org.cru.godtools.model.Tool;
@@ -54,7 +55,6 @@ import org.cru.godtools.tract.R;
 import org.cru.godtools.tract.R2;
 import org.cru.godtools.tract.adapter.ManifestPagerAdapter;
 import org.cru.godtools.tract.analytics.model.TractPageAnalyticsScreenEvent;
-import org.cru.godtools.tract.util.DrawableUtils;
 import org.cru.godtools.tract.util.ViewUtils;
 import org.cru.godtools.xml.content.ManifestLoader;
 import org.cru.godtools.xml.model.Card;
