+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/viewmodel/CallToActionViewHolder.java
@@ -8,9 +8,9 @@ import android.widget.ImageView;
 import android.widget.TextView;
 
 import org.cru.godtools.base.tool.model.view.TextViewUtils;
+import org.cru.godtools.base.ui.util.DrawableUtils;
 import org.cru.godtools.tract.R;
 import org.cru.godtools.tract.R2;
-import org.cru.godtools.tract.util.DrawableUtils;
 import org.cru.godtools.xml.model.CallToAction;
 
 import butterknife.BindView;
